Force crossing code to use doubles instead of doubles == int, which can cause issues as it would force the double to an int, breaking the crossing code as pointed out by Enrico.

+++ OpenSim/Region/Framework/Scenes/ScenePresence.cs
@@ -2422,10 +2422,10 @@ namespace OpenSim.Region.Framework.Scenes
 
                         foreach (GridRegion region in neighbors)
                         {
-                        if (TargetX >= region.RegionLocX
-                            && TargetY >= region.RegionLocY
-                            && TargetX < region.RegionLocX + region.RegionSizeX
-                            && TargetY < region.RegionLocY + region.RegionSizeY)
+                            if (TargetX >= (double)region.RegionLocX
+                                && TargetY >= (double)region.RegionLocY
+                                && TargetX < (double)(region.RegionLocX + region.RegionSizeX)
+                                && TargetY < (double)(region.RegionLocY + region.RegionSizeY))
                             {
                                 neighborRegion = region;
                                 break;
